Fix OS auth client lint warnings
