+++ apps/os/src/auth/client.tsx
@@ -1,24 +1,8 @@
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-  type ReactNode,
-} from "react";
+import { useCallback, useEffect, useMemo, useState, type ReactNode } from "react";
 import { createIterateAuthClient, type SessionResponse } from "@iterate-com/auth/client";
-
-type AuthClientContextValue = {
-  session: SessionResponse | null;
-  loading: boolean;
-  signIn: () => void;
-  signOut: () => Promise<void>;
-  refresh: () => Promise<void>;
-};
+import { AuthClientContext, type AuthClientContextValue } from "~/auth/client-context.ts";
 
 const authClient = createIterateAuthClient();
-const AuthClientContext = createContext<AuthClientContextValue | null>(null);
 
 export function AuthClientProvider({ children }: { children: ReactNode }) {
   const [session, setSession] = useState<SessionResponse | null>(null);
@@ -53,12 +37,4 @@ export function AuthClientProvider({ children }: { children: ReactNode }) {
   );
 
   return <AuthClientContext.Provider value={value}>{children}</AuthClientContext.Provider>;
-}
-
-export function useAuthClient() {
-  const value = useContext(AuthClientContext);
-  if (!value) {
-    throw new Error("useAuthClient must be used within AuthClientProvider.");
-  }
-  return value;
 }
